perf: 优化租户组织机构数据权限相关问题

+++ opsli-modulars/opsli-modulars-system/src/main/java/org/opsli/modulars/system/role/service/impl/RoleServiceImpl.java
@@ -15,7 +15,9 @@
  */
 package org.opsli.modulars.system.role.service.impl;
 
+import cn.hutool.core.collection.CollUtil;
 import cn.hutool.core.convert.Convert;
+import cn.hutool.core.util.ArrayUtil;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.google.common.collect.Lists;
 import org.apache.commons.lang3.StringUtils;
@@ -25,6 +27,7 @@ import org.opsli.common.constants.MyBatisConstants;
 import org.opsli.common.exception.ServiceException;
 import org.opsli.common.utils.FieldUtil;
 import org.opsli.core.base.service.impl.CrudServiceImpl;
+import org.opsli.core.msg.CoreMsg;
 import org.opsli.core.persistence.querybuilder.GenQueryBuilder;
 import org.opsli.core.persistence.querybuilder.QueryBuilder;
 import org.opsli.core.persistence.querybuilder.chain.QueryDataPermsHandler;
@@ -35,6 +38,7 @@ import org.opsli.modulars.system.role.entity.SysRole;
 import org.opsli.modulars.system.role.mapper.RoleMapper;
 import org.opsli.modulars.system.role.service.IRoleMenuRefService;
 import org.opsli.modulars.system.role.service.IRoleService;
+import org.opsli.modulars.system.user.service.IUserRoleRefService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -56,6 +60,8 @@ public class RoleServiceImpl extends CrudServiceImpl<RoleMapper, SysRole, RoleMo
     private RoleMapper mapper;
     @Autowired
     private IRoleMenuRefService iRoleMenuRefService;
+    @Autowired
+    private IUserRoleRefService iUserRoleRefService;
 
     @Override
     @Transactional(rollbackFor = Exception.class)
@@ -106,39 +112,68 @@ public class RoleServiceImpl extends CrudServiceImpl<RoleMapper, SysRole, RoleMo
             throw new ServiceException(SystemMsg.EXCEPTION_ROLE_UNIQUE);
         }
 
-        return super.update(model);
+        model = super.update(model);
+        // 清除缓存
+        if(null != model){
+            clearCache(Convert.toStrArray(model.getId()));
+        }
+        return model;
     }
 
     @Override
+    @Transactional(rollbackFor = Exception.class)
     public boolean delete(String id) {
+        boolean roleUsed = iUserRoleRefService.isRoleUsed(id);
+        if(roleUsed){
+            // 角色删除失败, 改角色正在被其他用户使用
+            throw new ServiceException(SystemMsg.EXCEPTION_ROLE_USED);
+        }
+
         // 删除角色下 权限
         iRoleMenuRefService.delPermsByRoleIds(Convert.toList(String.class, id));
         return super.delete(id);
     }
 
     @Override
+    @Transactional(rollbackFor = Exception.class)
     public boolean deleteAll(String[] ids) {
+        boolean roleUsed = iUserRoleRefService.isRoleUsed(ids);
+        if(roleUsed){
+            // 角色删除失败, 改角色正在被其他用户使用
+            throw new ServiceException(SystemMsg.EXCEPTION_ROLE_USED);
+        }
+
         // 删除角色下 权限
         iRoleMenuRefService.delPermsByRoleIds(Convert.toList(String.class, ids));
         return super.deleteAll(ids);
     }
 
     @Override
+    @Transactional(rollbackFor = Exception.class)
     public boolean delete(RoleModel model) {
         if(model == null){
             return false;
         }
+
+        boolean roleUsed = iUserRoleRefService.isRoleUsed(model.getId());
+        if(roleUsed){
+            // 角色删除失败, 改角色正在被其他用户使用
+            throw new ServiceException(SystemMsg.EXCEPTION_ROLE_USED);
+        }
+
         // 删除角色下 权限
         iRoleMenuRefService.delPermsByRoleIds(Convert.toList(String.class, model.getId()));
         return super.delete(model);
     }
 
     @Override
+    @Transactional(rollbackFor = Exception.class)
     public boolean deleteAll(Collection<RoleModel> models) {
         List<String> roleIds = Lists.newArrayList();
         for (RoleModel model : models) {
             roleIds.add(model.getId());
         }
+
         // 删除角色下 权限
         iRoleMenuRefService.delPermsByRoleIds(roleIds);
         return super.deleteAll(models);
@@ -214,6 +249,43 @@ public class RoleServiceImpl extends CrudServiceImpl<RoleMapper, SysRole, RoleMo
     }
 
 
+    /**
+     * 清除缓存
+     * @param roleIds 角色ID
+     */
+    private void clearCache(String[] roleIds){
+        // 清空该角色下 用户缓存
+        List<String> userIdList = iUserRoleRefService.getUserIdListByRoleIds(roleIds);
+        if(CollUtil.isNotEmpty(userIdList)){
+            int cacheCount = 0;
+            for (String userId : userIdList) {
+                cacheCount += 4;
+                boolean tmp;
+                // 清空当期用户缓存角色、权限、菜单
+                tmp = UserUtil.refreshUserRoles(userId);
+                if(tmp){
+                    cacheCount--;
+                }
+                tmp = UserUtil.refreshUserAllPerms(userId);
+                if(tmp){
+                    cacheCount--;
+                }
+                tmp = UserUtil.refreshUserMenus(userId);
+                if(tmp){
+                    cacheCount--;
+                }
+                tmp = UserUtil.refreshUserDefRole(userId);
+                if(tmp){
+                    cacheCount--;
+                }
+            }
+            // 判断删除状态
+            if(cacheCount != 0){
+                // 删除缓存失败
+                throw new ServiceException(CoreMsg.CACHE_DEL_EXCEPTION);
+            }
+        }
+    }
 
 }
